Switch to .gen file
