+++ scripts/generate-rtk-apis.ts
@@ -38,18 +38,12 @@ const config: ConfigFile = {
     },
     '../public/app/features/provisioning/api/endpoints.gen.ts': {
       apiFile: '../public/app/features/provisioning/api/baseAPI.ts',
-      schemaFile: '../pkg/tests/apis/openapi_snapshots/provisioning.grafana.app-v0alpha1.json',
+      schemaFile: '../data/openapi/provisioning.grafana.app-v0alpha1.json',
       apiImport: 'baseAPI',
       filterEndpoints,
       argSuffix: 'Arg',
       responseSuffix: 'Response',
       tag: true,
-      endpointOverrides: [
-        {
-          pattern: /^list/,
-          parameterFilter: () => false,
-        },
-      ],
     },
   },
 };
